Create manage_patchset_v2.sh
This script will ask for patchset, feeds to be updated, and which options for updating feed, as well as which tools, KernelCare, LibCare, Qemu. which are optional as it will default to kernel if field left blank. 

Edit script to put in your ePortal Base Url and your API key.